+++ setup.py
@@ -54,7 +54,7 @@ setup(
     packages=packages,
     entry_points={
         'console_scripts': [
-            'xcube = xcube.cli:main',
+            'xcube = xcube.cli.cli:main',
             'xcube-genl2c = xcube.genl2c.cli:main',
             'xcube-genl3 = xcube.genl3.cli:main',
             'xcube-grid = xcube.grid.main:main',
