+++ xcube/grid/main.py
@@ -19,7 +19,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from xcube.grid.cli import cli
+from xcube.cli.grid import cli
 
 
 def main(args=None):
